[FLINK-17182][network][tests] Fix the unstable RemoteInputChannelTest.testConcurrentOnSenderBacklogAndRecycle
In this unstable unit test, the exclusive buffers and floating buffers are recycled by different
threads, which might cause unexpected race condition issue. But actually they should always be
recycled by the same task thread in practice. So we refactor the test process to recycle them in
the same thread to avoid potential unnecessary issues.
    
This closes #11924.